--- a/tests/test_plotting.py
+++ b/tests/test_plotting.py
@@ -2,6 +2,7 @@ import pathlib
 import os
 import sys
 from weakref import proxy
+from pathlib import Path
 
 import imageio
 import numpy as np
@@ -521,9 +522,12 @@ def test_screenshot(tmpdir):
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")
 @pytest.mark.parametrize('ext', SUPPORTED_FORMATS)
 def test_save_screenshot(tmpdir, sphere, ext):
+    filename = str(tmpdir.mkdir("tmpdir").join('tmp' + ext))
     plotter = pyvista.Plotter(off_screen=OFF_SCREEN)
     plotter.add_mesh(sphere)
-    plotter.screenshot(str(tmpdir.mkdir("tmpdir").join('tmp' + ext)))
+    plotter.screenshot(filename)
+    assert os.path.isfile(filename)
+    assert Path(filename).stat().st_size
 
 
 @pytest.mark.skipif(NO_PLOTTING, reason="Requires system to support plotting")