### Overview
Simple fix to speedup screenshots (at least on Linux).  Discovered this while working on potentially automating tests using screenshots.

### Details
Here's the screenshot speed on Linux and Windows

```py
import pyvista as pv

plotter = pv.Plotter(off_screen=True)
plotter.add_mesh(pv.Sphere(), smooth_shading=True)
timeit plotter.screenshot('/tmp/tmp.png')
```

#### `feat/faster_screenshot` (this branch)

```
# LINUX:   70.9 ms ± 212 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# WINDOWS: 70 ms ± 145 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

#### `master`
```
# LINUX:   324 ms ± 564 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
# WINDOWS: 312 ms ± 462 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### MISC
I've also made a minor change in the tests to make sure we test all file types.